Fixed the map bug for real and some patch stuff
Map bug 2: electric boogaloo
--------------------------------
Setting to local directory created an issue where multiple presses of the setup button would try to go deeper in that directory (the-image-place/the-image-place/the-image-place ad infinitum). Resolved by deleting the set-directory command entirely. Make sure you keep the AOO2 file in the same directory as the nlogo file so this will work.

Patch stuff
-----------
Created an experimental go button- pretty sure this already exists in David's branch but I'm not seeing that on the git repository so I made one to test patch colors. Invasion code in progress.